Update 2
I have completed the game of life!
... with only a couple bugs. The top and bottom rows do not work, everything will die there.
also, when a cell goes off the screen to the right or left it will appear on the other side, except one row down (if its going to the right)
or one row up (if it's going to the left).
I also added buttons to show continuous generations and to stop them.
I will fix these bugs when I can!